+++ src/server/serverManager.ts
@@ -9,6 +9,7 @@ import {
 	AutoRefreshPreview,
 	SettingUtil,
 	Settings,
+	SETTINGS_SECTION_ID,
 } from '../utils/settingsUtil';
 import {
 	DONT_SHOW_AGAIN,
@@ -19,12 +20,39 @@ import TelemetryReporter from 'vscode-extension-telemetry';
 import { EndpointManager } from '../infoManagers/endpointManager';
 // import { WorkspaceManager } from '../infoManagers/workspaceManager';
 import { ConnectionManager } from '../connectionInfo/connectionManager';
-import { serverMsg } from '../serverGrouping';
 import { PathUtil } from '../utils/pathUtil';
 import { Connection } from '../connectionInfo/connection';
+import { PreviewManager } from '../editorPreview/previewManager';
+import { ServerStartedStatus, ServerTaskProvider } from '../task/serverTaskProvider';
+
+/**
+ * @description the server log item that is sent from the HTTP server to the server logging task.
+ */
+ export interface serverMsg {
+	method: string;
+	url: string;
+	status: number;
+}
+
+/**
+ * @description the info for launching a preview, used after a server is launched.
+ */
+export interface launchInfo {
+	external: boolean;
+	file: string;
+	relative: boolean;
+	debug: boolean;
+	panel?: vscode.WebviewPanel;
+	connection: Connection;
+}
 
 const localize = nls.loadMessageBundle();
 export class ServerManager extends Disposable {
+	private readonly _onClose = this._register(
+		new vscode.EventEmitter<void>()
+	);
+	public readonly onClose = this._onClose.event;
+	private _pendingLaunchInfo: launchInfo | undefined;
 	private readonly _httpServer: HttpServer;
 	private readonly _wsServer: WSServer;
 	private _isServerOn = false;
@@ -32,13 +60,18 @@ export class ServerManager extends Disposable {
 	private _httpConnected = false;
 	private readonly _watcher;
 
+	public get port(): number | undefined {
+		return this._connection.httpPort;
+	}
+
 	constructor(
 		private readonly _extensionUri: vscode.Uri,
 		_reporter: TelemetryReporter,
 		_endpointManager: EndpointManager,
-		// _workspaceManager: WorkspaceManager,
 		private readonly _connection: Connection,
 		private readonly _statusBar: StatusBarNotifier,
+		private readonly _previewManager: PreviewManager,
+				private readonly _serverTaskProvider: ServerTaskProvider,
 		_userDataDir: string | undefined
 	) {
 		super();
@@ -132,10 +165,57 @@ export class ServerManager extends Disposable {
 				}
 			})
 		);
+		this._connection.onConnected((e) => {
+			this._serverTaskProvider.serverStarted(
+				e.httpURI,
+				ServerStartedStatus.JUST_STARTED
+			);
 
+			if (this._pendingLaunchInfo) {
+				if (this._pendingLaunchInfo.external) {
+					this._previewManager.launchFileInExternalBrowser(
+						this._pendingLaunchInfo.file,
+						this._pendingLaunchInfo.relative,
+						this._pendingLaunchInfo.debug,
+						this._connection
+					);
+				} else {
+					this._previewManager.launchFileInEmbeddedPreview(
+						this._pendingLaunchInfo.file,
+						this._pendingLaunchInfo.relative,
+						this._pendingLaunchInfo.panel,
+						this._connection
+					);
+				}
+
+				this._pendingLaunchInfo = undefined;
+			}
+		});
+
+		vscode.workspace.onDidChangeConfiguration((e) => {
+			if (e.affectsConfiguration(SETTINGS_SECTION_ID)) {
+				this.updateConfigurations();
+				this._connection.pendingPort = SettingUtil.GetConfig(
+					this._extensionUri
+				).portNumber;
+				this._connection.pendingHost = SettingUtil.GetConfig(
+					this._extensionUri
+				).hostIP;
+			}
+		});
+
+		this._connection.onConnected((e) => {
+			this._serverTaskProvider.serverStarted(
+				e.httpURI,
+				ServerStartedStatus.JUST_STARTED
+			);
+		});
 		vscode.commands.executeCommand('setContext', LIVE_PREVIEW_SERVER_ON, false);
 	}
 
+	public get workspace(): vscode.WorkspaceFolder | undefined {
+		return this._connection.workspace;
+	}
 	/**
 	 * @returns {boolean} whether the HTTP server is on.
 	 */
@@ -161,13 +241,29 @@ export class ServerManager extends Disposable {
 	/**
 	 * @description close the server instances.
 	 */
-	public closeServer(): void {
-		this._statusBar.RemoveServer(this._connection.workspace?.uri);
-		this._httpServer.close();
-		this._wsServer.close();
-		this._isServerOn = false;
-		this.showServerStatusMessage('Server Stopped');
-		vscode.commands.executeCommand('setContext', LIVE_PREVIEW_SERVER_ON, false);
+	public closeServer(): boolean {
+		if (this.isRunning) {
+			this._statusBar.RemoveServer(this._connection.workspace?.uri);
+			this._httpServer.close();
+			this._wsServer.close();
+			this._isServerOn = false;
+			this.showServerStatusMessage('Server Stopped');
+			this._onClose.fire();
+			if (
+				this._previewManager.currentPanel &&
+				this._previewManager.currentPanel.currentConnection === this._connection
+			) {
+				this._previewManager.currentPanel?.close();
+			}
+
+			if (this._serverTaskProvider.isRunning) {
+				this._serverTaskProvider.serverStop(true);
+			}
+
+			this._connection.disconnected();
+			return true;
+		}
+		return false;
 	}
 
 	/**
@@ -175,7 +271,11 @@ export class ServerManager extends Disposable {
 	 * @param {number} port the port to try to start the HTTP server on.
 	 * @returns {boolean} whether the server has been started correctly.
 	 */
-	public openServer(port: number): boolean {
+	public openServer(fromTask = false): boolean {
+
+		const port = this._connection.httpPort;
+		if (!this.isRunning) {
+
 		this._httpConnected = false;
 		this._wsConnected = false;
 		if (this._extensionUri) {
@@ -186,6 +286,16 @@ export class ServerManager extends Disposable {
 			return true;
 		}
 		return false;
+		} else if (fromTask) {
+			this._connection.resolveExternalHTTPUri().then((uri) => {
+				this._serverTaskProvider.serverStarted(
+					uri,
+					ServerStartedStatus.STARTED_BY_EMBEDDED_PREV
+				);
+			});
+		}
+
+		return true;
 	}
 
 	/**
@@ -279,5 +389,114 @@ export class ServerManager extends Disposable {
 					}
 				});
 		}
+	}
+
+
+
+	dispose() {
+		this.closeServer();
+	}
+
+	/**
+	 * Opens the preview in an external browser.
+	 * @param {string} file the filesystem path to open in the preview.
+	 * @param {boolean} relative whether the path was absolute or relative to the current workspace.
+	 * @param {boolean} debug whether or not to run in debug mode.
+	 */
+	public showPreviewInBrowser(
+		file = '/',
+		relative = true,
+		debug = false
+	): void {
+		if (!this._serverTaskProvider.isRunning) {
+			if (!this.isRunning) {
+				// set the pending launch info, which will trigger once the server starts in `launchFileInExternalPreview`
+				this._pendingLaunchInfo = {
+					external: true,
+					file: file,
+					relative: relative,
+					debug: debug,
+					connection: this._connection,
+				};
+			} else {
+				this._previewManager.launchFileInExternalBrowser(
+					file,
+					relative,
+					debug,
+					this._connection
+				);
+			}
+			if (
+				vscode.workspace.workspaceFolders &&
+				vscode.workspace.workspaceFolders.length > 0 &&
+				this._previewManager.runTaskWithExternalPreview
+			) {
+				this._serverTaskProvider.extRunTask(
+					SettingUtil.GetConfig(this._extensionUri)
+						.browserPreviewLaunchServerLogging
+				);
+			} else {
+				// global tasks are currently not supported, just turn on server in this case.
+				const serverOn = this.openServer();
+
+				if (!serverOn) {
+					return;
+				}
+			}
+		} else {
+			this._previewManager.launchFileInExternalBrowser(
+				file,
+				relative,
+				debug,
+				this._connection
+			);
+		}
+	}
+
+
+	public get running() {
+		return this.isRunning;
+	}
+
+	/**
+	 * Creates an (or shows the existing) embedded preview.
+	 * @param {vscode.WebviewPanel} panel the panel, which may have been serialized from a previous session.
+	 * @param {string} file the filesystem path to open in the preview.
+	 * @param {boolean} relative whether the path was absolute or relative to the current workspace.
+	 * @param {boolean} debug whether to run in debug mode (not implemented).
+	 */
+	public createOrShowEmbeddedPreview(
+		panel: vscode.WebviewPanel | undefined = undefined,
+		file = '/',
+		relative = true,
+		debug = false
+	): void {
+		if (!this.isRunning) {
+			// set the pending launch info, which will trigger once the server starts in `launchFileInEmbeddedPreview`
+			this._pendingLaunchInfo = {
+				external: false,
+				panel: panel,
+				file: file,
+				relative: relative,
+				debug: debug,
+				connection: this._connection,
+			};
+			this.openServer();
+		} else {
+			this._previewManager.launchFileInEmbeddedPreview(
+				file,
+				relative,
+				panel,
+				this._connection
+			);
+		}
+	}
+
+	/**
+	 * @param {string} file the path to test.
+	 * @returns {boolean} whether the path exists when placed relative to the workspae root.
+	 */
+	public pathExistsRelativeToWorkspace(file: string): boolean {
+		return this._connection.pathExistsRelativeToWorkspace(file);
 	}
 }
